+++ ext/MRISubspaceReconCUDAExt/NFFTNormalOp.jl
@@ -4,12 +4,17 @@ function MRISubspaceRecon.NFFTNormalOp(
     U::AnyCuArray{Tc};
     cmaps=(1,),
     sample_mask=CUDA.ones(Bool, size(trj)[2:end]),
-    verbose=false
+    verbose=false,
+    lowmem=false,
     ) where {T <: Real, Tc <: Union{T, Complex{T}}}
 
-    Λ, kmask_indcs = calculate_kernel_noncartesian(2 .* img_shape, trj, U; sample_mask, verbose)
-
-    return MRISubspaceRecon.NFFTNormalOp(img_shape, Λ, kmask_indcs; cmaps=cmaps)
+    if lowmem
+        Λ_d, kmask_d = MRISubspaceRecon.calculate_kernel_lowmem(img_shape, trj, U; sample_mask, verbose)
+        return MRISubspaceRecon.NFFTNormalOpLowmem(img_shape, Λ_d, kmask_d; cmaps)
+    else
+        Λ, kmask_indcs = calculate_kernel_noncartesian(2 .* img_shape, trj, U; sample_mask, verbose)
+        return MRISubspaceRecon.NFFTNormalOp(img_shape, Λ, kmask_indcs; cmaps=cmaps)
+    end
 end
 
 # Wrapper for 4D data arrays